LOG4J2-2011 moved classes from util.picocli to tools.picocli
